refactor(json-crdt): 💡 move NodeEvents one folder up
